MAINT: Generate metadata from _metadata
